+++ src/utils/hello-world.ts
@@ -1,9 +1,0 @@
-/**
- * Return a personalized greeting.
- *
- * @param name - person's name to greet
- * @returns greeting
- */
-export function sendHello(name: string): string {
-  return `Hello ${name}`;
-}