+++ CHANGELOG.md
@@ -21,6 +21,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [0.2.0] - 2024-08-20
+
+### Added
+
+- Added initial WIP content including unit tests:
+  - base model (minimal implementation)
+  - core models based on the [FHIR Type Framework](https://hl7.org/fhir/R5/types.html)
+  - Extension model
+  - primitive type definitions
+  - selected primitive datatype models
+  - Period complex data type model
+
+### Changed
+
+- Updated dependencies
+- Minor updates to various project configurations
+
+### Removed
+
+- Removed the source and tests from the `typescript-template` project
+
 ## [0.1.0] - 2024-07-19
 
 ### Added